users_api/pyproject.toml
[tool.poetry]
name = "microservices"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "src"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.120.4"
uvicorn = "^0.38.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
pydantic-settings = "^2.1.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.3.0"
email-validator = "^2.1.0"
alembic = "^1.13.0"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
isort = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
